[Feat/#17] DTO, Entity 분리 등
