+++ GLINT-iOS/GLINT-iOS/Source/Domain/UseCase/DetailViewUseCase.swift
@@ -10,6 +10,6 @@ import Foundation
 struct DetailViewUseCase {
     var filterDetail: @Sendable (_ filterID: String) async throws -> FilterDetailEntity
     var createOrder: @Sendable (_ request: CreateOrderEntity.Request) async throws -> CreateOrderEntity.Response
-    var infoOrder: @Sendable () async throws -> InfoOrderEntity.Response
+    var infoOrder: @Sendable () async throws -> OrderHistoryEntity.Response
     var paymentValidation: @Sendable (_ request: PaymentValidationEntity.Request) async throws -> PaymentValidationEntity.Response
 }
